added weapons to characters

+++ Services/CharacterServices/CharacterService.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using AutoMapper;
 using Microsoft.EntityFrameworkCore;
@@ -18,11 +19,18 @@ namespace WEB_API_Udemy.Services.CharacterServices
         };
         private readonly IMapper _mapper;
         private readonly DataContext _context;
+        private readonly IHttpContextAccessor _httpContextAccessor;
 
-        public CharacterService(IMapper mapper, DataContext context)
+        public CharacterService(IMapper mapper, DataContext context, IHttpContextAccessor httpContextAccessor)
         {
             _mapper = mapper;
             _context = context;
+            _httpContextAccessor = httpContextAccessor;
+        }
+
+        private int GetUserId()
+        {
+           return int.Parse(_httpContextAccessor.HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier));
         }
 
      
@@ -30,9 +38,13 @@ namespace WEB_API_Udemy.Services.CharacterServices
         {
             var serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
             Character character = _mapper.Map<Character>(newCharacter);
+
+            character.User = await _context.Users.FirstOrDefaultAsync(u => u.Id == GetUserId());
+
             _context.Characters.Add(character);
             await _context.SaveChangesAsync();
             serviceResponse.Data = await _context.Characters
+                .Where(c => c.User.Id == GetUserId())
                 .Select(c => _mapper.Map<GetCharacterDto>(c))
                 .ToListAsync();
             return serviceResponse;
@@ -44,12 +56,21 @@ namespace WEB_API_Udemy.Services.CharacterServices
 
             try
             {
-                Character character = await _context.Characters.FirstAsync(c => c.Id == id);
-
-                _context.Characters.Remove(character);
-                await _context.SaveChangesAsync();
-
-                serviceResponse.Data = _context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
+                Character character = await _context.Characters
+                    .FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
+                    if (character != null)
+                    {
+                        _context.Characters.Remove(character);
+                        await _context.SaveChangesAsync();
+                        serviceResponse.Data = _context.Characters
+                            .Where(c => c.User.Id == GetUserId())
+                            .Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
+                    }
+                    else
+                    {
+                        serviceResponse.Success = false;
+                        serviceResponse.Message = "Character Not Found";
+                    }   
                 
             } 
             catch (Exception ex)
@@ -63,7 +84,7 @@ namespace WEB_API_Udemy.Services.CharacterServices
         public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters()
         {
             var response = new ServiceResponse<List<GetCharacterDto>>();
-            var dbCharacters = await _context.Characters.ToListAsync();
+            var dbCharacters = await _context.Characters.Where(c => c.User.Id == GetUserId()).ToListAsync();
             response.Data = dbCharacters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
 
             return response;
@@ -72,7 +93,8 @@ namespace WEB_API_Udemy.Services.CharacterServices
         public async Task<ServiceResponse<GetCharacterDto>> GetCharacterById(int id)
         {
             var serviceResponse = new ServiceResponse<GetCharacterDto>();
-            var dbCharacter = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id); 
+            var dbCharacter = await _context.Characters
+                .FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId()); 
             serviceResponse.Data = _mapper.Map<GetCharacterDto>(dbCharacter);
 
             return serviceResponse; 
@@ -83,21 +105,33 @@ namespace WEB_API_Udemy.Services.CharacterServices
            var serviceResponse = new ServiceResponse<GetCharacterDto>();
 
            try{
-           var character = await _context.Characters.FirstOrDefaultAsync(c => c.Id == updatedCharacter.Id);
+           var character = await _context.Characters.Include(c => c.User).FirstOrDefaultAsync(c => c.Id == updatedCharacter.Id);
+
+           if (character.User.Id == GetUserId())
+           {
+                character.Name = updatedCharacter.Name;
+                character.HitPoints = updatedCharacter.HitPoints;
+                character.Strength = updatedCharacter.Strength;
+                character.Defense = updatedCharacter.Defense;
+                character.Class = updatedCharacter.Class;
+
+                await _context.SaveChangesAsync();
+
+                serviceResponse.Data = _mapper.Map<GetCharacterDto>(character);
+           }
+           else 
+           {
+            serviceResponse.Success = false;
+            serviceResponse.Message = "Character Not Found";
+           }
 
             // creates a new character with the values provided, if vbalues are missing it uses the default class values and fills in the rest
 
             // This code take in the new values and updates them.  The unchaged values remain the same and the character continues.
 
-           character.Name = updatedCharacter.Name;
-           character.HitPoints = updatedCharacter.HitPoints;
-           character.Strength = updatedCharacter.Strength;
-           character.Defense = updatedCharacter.Defense;
-           character.Class = updatedCharacter.Class;
+           
 
-           await _context.SaveChangesAsync();
-
-           serviceResponse.Data = _mapper.Map<GetCharacterDto>(character);
+           
            } catch(Exception ex){
             serviceResponse.Success = false;
             serviceResponse.Message = ex.Message;
